refactor: change redax store types and import
