+++ src/components/Header/AuthIcon/AuthIcon.tsx
@@ -2,7 +2,7 @@ import cn from 'classnames';
 import styles from './AuthIcon.module.css';
 import { Link } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import { StoreType } from '../../../index';
+import { StoreType } from '../../../store/store';
 
 const AuthIcon: React.FC = () => {
   const isLogin = useSelector((state: StoreType): boolean => state.auth.isLogin);
