+++ src/hooks/redax-hooks.ts
@@ -1,0 +1,5 @@
+import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
+import { AppDispatchType, StoreType } from "../store/store";
+
+export const useCustomDispatch = () => useDispatch<AppDispatchType>();
+export const useCustomSelector: TypedUseSelectorHook<StoreType> = useSelector;
